[9.0][WIP] connector_prestashop_features: New module
