+++ connector_prestashop/models/product_product/importer.py
@@ -4,7 +4,11 @@
 from openerp import models
 
 from openerp.addons.connector.unit.backend_adapter import BackendAdapter
-from openerp.addons.connector.unit.mapper import mapping, ImportMapper
+from openerp.addons.connector.unit.mapper import (
+    mapping,
+    ImportMapper,
+    only_create
+)
 from ...unit.importer import (
     PrestashopImporter,
     import_batch,
@@ -247,6 +251,16 @@ class ProductCombinationMapper(ImportMapper):
             'impact_price': impact
         }
 
+    @only_create
+    @mapping
+    def odoo_id(self, record):
+        Propuct = self.env['product.product']
+        product = Propuct.search([
+            ('name', '=', record['name'])
+        ])
+        if Propuct:
+            return {'odoo_id': product.id}
+
 
 @prestashop
 class ProductCombinationOptionImporter(PrestashopImporter):
@@ -318,6 +332,16 @@ class ProductCombinationOptionMapper(ImportMapper):
         else:
             name = record['name']
         return {'name': name}
+    
+    @only_create
+    @mapping
+    def odoo_id(self, record):
+        ProductAttribute = self.env['product.attribute']
+        product_attribute = ProductAttribute.search([
+            ('name', '=', record['name'])
+        ])
+        if product_attribute:
+            return {'odoo_id': product_attribute.id}
 
 
 @prestashop
@@ -353,6 +377,16 @@ class ProductCombinationOptionValueMapper(ImportMapper):
     @mapping
     def backend_id(self, record):
         return {'backend_id': self.backend_record.id}
+    
+    @only_create
+    @mapping
+    def odoo_id(self, record):
+        AttributeValue = self.env['product.attribute.value']
+        attribute_value = AttributeValue.search([
+            ('name', '=', record['name'])
+        ])
+        if attribute_value:
+            return {'odoo_id': attribute_value.id}
 
 
 @prestashop
