+++ connector_prestashop_catalog_feature/__init__.py
@@ -4,4 +4,3 @@
 from . import consumer
 from . import models
 from . import wizard
-
