+++ connector_prestashop_catalog_feature/__openerp__.py
@@ -15,13 +15,9 @@
               'Odoo Community Association (OCA)',
     "category": "Connector",
     "data": [
-        # 'views/product_attribute_view.xml',
-        # 'views/product_view.xml',
-        # 'wizard/export_category_view.xml',
-        # 'wizard/export_multiple_products_view.xml',
-        # 'wizard/sync_products_view.xml',
-        # 'wizard/active_deactive_products_view.xml',
-        # 'views/product_image_view.xml',
+        'views/custom_info_property_view.xml',
+        'views/custom_info_option_view.xml',
+        'wizard/export_feature_view.xml',
     ],
     "auto_install": True,
     "installable": True,
