--- connector_prestashop_catalog_feature/views/product_view.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record id="prestashop_product_product_form" model="ir.ui.view">
-            <field name="name">connector_prestashop.product.template.form
-            </field>
-            <field name="model">prestashop.product.template</field>
-            <field name="inherit_id"
-                   ref="connector_prestashop.product_connector_presta_form_view"/>
-            <field name="type">form</field>
-            <field name="arch" type="xml">
-                <xpath expr="/form/group[2]" position="inside">
-                    <group col="4" colspan="4">
-                        <field name="meta_title"/>
-                        <field name="meta_description"/>
-                        <field name="meta_keywords"/>
-                        <field name="tags"/>
-                        <field name="online_only"/>
-                        <field name="additional_shipping_cost"/>
-                        <field name="available_now"/>
-                        <field name="available_later"/>
-                        <field name="available_date"/>
-                        <field name="minimal_quantity"/>
-                    </group>
-                </xpath>
-            </field>
-        </record>
-    </data>
-</openerp>
