--- connector_prestashop_catalog_feature/wizard/active_deactive_products_view.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record model="ir.ui.view" id="active_product_form">
-            <field name="name">active.product.form</field>
-            <field name="model">active.deactive.products</field>
-            <field name="arch" type="xml">
-                <form string="Export Products">
-                    <group>
-                        <label string="Active selected products"/>
-                    </group>
-                    <group>
-                        <field name="force_status"/>
-                    </group>
-                    <footer>
-                        <button name="active_products" type="object"
-                                string="Active" class="oe_highlight"/>
-                        or
-                        <button string="Cancel" class="oe_link"
-                                special="cancel"/>
-                    </footer>
-                </form>
-            </field>
-        </record>
-        <act_window name="Active Products"
-                    res_model="active.deactive.products"
-                    src_model="product.template"
-                    view_mode="form"
-                    target="new"
-                    view_id="active_product_form"
-                    key2="client_action_multi"
-                    id="act_active_products"/>
-                    
-        <record model="ir.ui.view" id="deactive_product_form">
-            <field name="name">deactive.product.form</field>
-            <field name="model">active.deactive.products</field>
-            <field name="arch" type="xml">
-                <form string="Export Products">
-                    <group>
-                        <label string="deactive selected products"/>
-                    </group>
-                    <footer>
-                        <button name="deactive_products" type="object"
-                                string="Deactive" class="oe_highlight"/>
-                        or
-                        <button string="Cancel" class="oe_link"
-                                special="cancel"/>
-                    </footer>
-                </form>
-            </field>
-        </record>
-        <act_window name="Deactive Products"
-                    res_model="active.deactive.products"
-                    src_model="product.template"
-                    view_mode="form"
-                    target="new"
-                    view_id="deactive_product_form"
-                    key2="client_action_multi"
-                    id="act_deactive_products"/>
-    </data>
-</openerp>
