+++ connector_prestashop_catalog_feature/wizard/export_feature_view.xml
@@ -2,17 +2,16 @@
 <openerp>
 <data>
 
-    <record id="export_category_form" model="ir.ui.view">
-        <field name="name">wiz.prestashop.export.category.form</field>
-        <field name="model">wiz.prestashop.export.category</field>
+    <record id="export_feature_form" model="ir.ui.view">
+        <field name="name">wiz.prestashop.export.feature.form</field>
+        <field name="model">wiz.prestashop.export.feature</field>
         <field name="arch" type="xml">
-            <form string="Export Categories">
+            <form string="Export Product Features">
                 <group>
                     <field name="backend_id"/>
-                    <field name="shop_id" domain="[('backend_id','=',name)]"/>
                 </group>
                 <footer>
-                    <button name="export_categories" type="object"
+                    <button name="export_features" type="object"
                             string="Export" class="oe_highlight"/>
                     or
                     <button string="Cancel" class="oe_link"
@@ -22,14 +21,14 @@
         </field>
     </record>
 
-    <act_window id="act_export_prestashop_category"
+    <act_window id="act_export_prestashop_features"
                 name="Export To PrestaShop"
-                res_model="wiz.prestashop.export.category"
-                src_model="product.category"
+                res_model="wiz.prestashop.export.feature"
+                src_model="custom.info.property"
                 view_mode="form"
                 target="new"
                 key2="client_action_multi"
-                view_id="export_category_form"/>
+                view_id="export_feature_form"/>
 
 </data>
 </openerp>
