--- connector_prestashop_catalog_feature/wizard/export_multiple_products_view.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record model="ir.ui.view" id="export_multiple_products_form">
-            <field name="name">export.multiple.products.form</field>
-            <field name="model">export.multiple.products</field>
-            <field name="arch" type="xml">
-                <form string="Export Products">
-                    <group>
-                        <field name="name"/>
-                        <field name="shop" domain="[('backend_id','=',name)]"/>
-                    </group>
-                    <footer>
-                        <button name="export_products" type="object"
-                                string="Export" class="oe_highlight"/>
-                        or
-                        <button string="Cancel" class="oe_link"
-                                special="cancel"/>
-                    </footer>
-                </form>
-            </field>
-        </record>
-        <act_window name="Export Products"
-                    res_model="export.multiple.products"
-                    src_model="product.template"
-                    view_mode="form"
-                    target="new"
-                    key2="client_action_multi"
-                    id="act_export_multiple_products"
-                    view_id="export_multiple_products_form"/>
-
-        <record model="ir.ui.view" id="export_variant_stock_form">
-            <field name="name">export.variant.stock.form</field>
-            <field name="model">export.multiple.products</field>
-            <field name="arch" type="xml">
-                <form string="Export variant stock">
-                    <footer>
-                        <button name="export_variant_stock" type="object"
-                                string="Export" class="oe_highlight"/>
-                        or
-                        <button string="Cancel" class="oe_link"
-                                special="cancel"/>
-                    </footer>
-                </form>
-            </field>
-        </record>
-        <act_window name="Export Products Stock"
-                    res_model="export.multiple.products"
-                    src_model="product.template"
-                    view_mode="form"
-                    target="new"
-                    key2="client_action_multi"
-                    id="act_export_products_stock"
-                    view_id="export_variant_stock_form"/>
-
-    </data>
-</openerp>
