--- connector_prestashop_catalog_feature/wizard/sync_products_view.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record model="ir.ui.view" id="sync_products_form">
-            <field name="name">sync.products.form</field>
-            <field name="model">sync.products</field>
-            <field name="arch" type="xml">
-                <form string="Export Products">
-                    <group>
-                        <label string="sync selected products"/>
-                    </group>
-                    <footer>
-                        <button name="sync_products" type="object"
-                                string="Sync" class="oe_highlight"/>
-                        or
-                        <button string="Cancel" class="oe_link"
-                                special="cancel"/>
-                    </footer>
-                </form>
-            </field>
-        </record>
-        <act_window name="Sync Products"
-                    res_model="sync.products"
-                    src_model="product.template"
-                    view_mode="form"
-                    target="new"
-                    key2="client_action_multi"
-                    id="act_presta_sync_products"/>
-    </data>
-</openerp>
