+++ connector_prestashop_catalog_manager/__init__.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import product_category
-from . import product
-from . import product_attribute
-from . import product_combination
+from . import consumer
+from . import models
 from . import wizard
-from . import product_image
