+++ connector_prestashop_catalog_manager/models/product_image/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import exporter
