+++ connector_prestashop_catalog_manager/models/product_product/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from . import exporter
+from . import common
