+++ connector_prestashop_catalog_manager/models/product_template/common.py
@@ -1,0 +1,46 @@
+# -*- coding: utf-8 -*-
+# © 2016 Sergio Teruel <sergio.teruel@tecnativa.com>
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from openerp import models, fields
+import openerp.addons.decimal_precision as dp
+
+
+class PrestashopProductTemplate(models.Model):
+    _inherit = 'prestashop.product.template'
+
+    meta_title = fields.Char(
+        string='Meta Title',
+        translate=True
+    )
+    meta_description = fields.Char(
+        string='Meta Description',
+        translate=True
+    )
+    meta_keywords = fields.Char(
+        string='Meta Keywords',
+        translate=True
+    )
+    tags = fields.Char(
+        string='Tags',
+        translate=True
+    )
+    online_only = fields.Boolean(string='Online Only')
+    additional_shipping_cost = fields.Float(
+        string='Additional Shipping Price',
+        digits_compute=dp.get_precision('Product Price'),
+        help="Additionnal Shipping Price for the product on Prestashop")
+    available_now = fields.Char(
+        string='Available Now',
+        translate=True
+    )
+    available_later = fields.Char(
+        string='Available Later',
+        translate=True
+    )
+    available_date = fields.Date(string='Available Date')
+    minimal_quantity = fields.Integer(
+        string='Minimal Quantity',
+        help='Minimal Sale quantity',
+        default=1,
+    )
