+++ connector_prestashop_feature/models/product_template/importer.py
@@ -5,8 +5,6 @@
 import logging
 from openerp.addons.connector.unit.mapper import (
     mapping,
-    backend_to_m2o,
-    ImportMapper,
 )
 from openerp.addons.connector_prestashop.backend import prestashop
 from openerp.addons.connector_prestashop.models.product_template import \
@@ -54,18 +52,18 @@ class TemplateRecordImportFeatures(importer.ProductTemplateImporter):
         CustomInfoValue = self.session.env['custom.info.value']
         property_binder = self.binder_for('prestashop.product.features')
         option_binder = self.binder_for('prestashop.product.feature.values')
-        
+
         info_value_domain = [
             ('model', '=', 'product.template'),
             ('res_id', '=', binding.odoo_id.id),
         ]
         initial_info_value = CustomInfoValue.search(info_value_domain)
-        
+
         properties = []
         for feature_value in feature_values:
             properties.append(
                 property_binder.to_odoo(feature_value['id'], unwrap=True).id)
-        
+
         if properties:
             info_value_domain.append(('property_id', 'in', properties))
         ps_info_values = CustomInfoValue.search(info_value_domain)
@@ -73,7 +71,7 @@ class TemplateRecordImportFeatures(importer.ProductTemplateImporter):
         if empty_property_value:
             empty_property_value.with_context(
                 connector_no_export=True).write({'value_id': False})
-       
+
         for feature_value in feature_values:
             info_property = property_binder.to_odoo(
                 feature_value['id'], unwrap=True)
